+++ src/main/java/com/hcmut/voltrent/service/payment/PaymentService.java
@@ -64,7 +64,7 @@ public class PaymentService implements IPaymentService {
         Payment payment = Payment.builder()
                 .gateway(paymentRequest.getGateway().getValue())
                 .bookingId(paymentRequest.getBookingId())
-                .totalAmount((long) paymentRequest.getTotalAmount())
+                .totalAmount(paymentRequest.getTotalAmount())
                 .transactionRef(paymentRequest.getTransactionId())
                 .status(paymentRequest.getPaymentStatus().getDescription())
                 .partnerPayDate(paymentRequest.getPartnerPayDate())
